Paginated declaration modal factored out
