+++ frontend/src/components/NewBepiasViews/DeclarationsTableSection/index.vue
@@ -34,20 +34,12 @@
           label="Télécharger"
           @click="exportModalOpened = true"
         />
-        <DsfrModal title="Exporter votre recherche" :opened="exportModalOpened" @close="exportModalOpened = false">
-          <p>
-            Votre recherche dépasse le seuil des {{ MAX_EXPORT_RESULTS }} lignes. Vous pouvez télécharger la totalité de
-            votre recherche en fichiers séparés.
-          </p>
-          <ul>
-            <li v-for="(link, index) in paginatedExcelLinks" :key="`download-links-${index}`">
-              {{ link.label }}
-              <a :href="link.url" download class="bg-none ml-2">
-                <DsfrButton label="Télécharger" tertiary no-outline size="sm" icon="ri-file-excel-2-fill"></DsfrButton>
-              </a>
-            </li>
-          </ul>
-        </DsfrModal>
+        <PaginatedExcelDownload
+          v-model="exportModalOpened"
+          :baseUrl="excelUrl"
+          :maxLines="MAX_EXPORT_RESULTS"
+          :totalLines="data.count"
+        />
       </div>
     </div>
     <!-- Zone des filtres actifs -->
@@ -157,6 +149,7 @@ import { toOptions } from "@/utils/forms.js"
 import ElementAutocomplete from "@/components/ElementAutocomplete.vue"
 import { typesMapping } from "@/utils/mappings"
 import IngredientFilterAccordeon from "./IngredientFilterAccordeon"
+import PaginatedExcelDownload from "@/components/PaginatedExcelDownload"
 
 const MAX_EXPORT_RESULTS = 3
 const exportModalOpened = ref(false)
@@ -353,32 +346,6 @@ const updateDoseStrings = (index, newValue) => {
 // Export Excel
 
 const excelUrl = computed(() => `/api/v1/control/declarations-export.xlsx?${commonApiParams.value}`)
-
-const paginatedExcelLinks = computed(() => {
-  const totalResults = data.value.results.length
-  if (!totalResults) return []
-
-  const links = []
-  const numFiles = Math.ceil(totalResults / MAX_EXPORT_RESULTS)
-
-  for (let i = 0; i < numFiles; i++) {
-    const offset = i * MAX_EXPORT_RESULTS
-    const limit = MAX_EXPORT_RESULTS
-    const pageNumber = i + 1
-    const startRange = offset + 1
-    const endRange = Math.min(offset + MAX_EXPORT_RESULTS, totalResults)
-
-    // Create the URL with pagination parameters
-    const url = `${excelUrl.value}&limit=${limit}&offset=${offset}`
-
-    links.push({
-      url: url,
-      label: `Lignes ${startRange} à ${endRange} (${pageNumber}/${numFiles})`,
-    })
-  }
-
-  return links
-})
 </script>
 
 <style scoped>
